[#85] Add issuesRowDifference function for testing DB updates better
Originally, to test if a table was updated I just made sure the
difference between the number of rows before and after the update was
0. This is not enough to test the table was actually updated.
issuesRowDifference allows the actual rows before and after the
update to be compared and ensure the table was updated. The difference
between the rows after the action and before the action should be equal
to the rows upserted, if all the rows were upserted successfully.

Resolves #85